Add externalHelpersPath configuration
I would like to be able to control the import path generated by this babel plugin.
The goal is generate code where the import path is the actual file url (not an alias that needs to be resolved by node).

To achieve this I would like to add "externalHelpersPath" configuration. 
What do you think about it?